add DI resolver for external authors
