+++ apps/crn-server/src/handlers/external-author/index-handler.ts
@@ -3,40 +3,15 @@ import {
   algoliaSearchClientFactory,
 } from '@asap-hub/algolia';
 import { EventBridgeHandler } from '@asap-hub/server-common';
-import {
-  RestExternalAuthor,
-  SquidexGraphql,
-  SquidexRest,
-} from '@asap-hub/squidex';
-import {
-  getGraphQLClient as getContentfulGraphQLClient,
-  getRestClient as getContentfulRestClient,
-} from '@asap-hub/contentful';
 import { EventBridgeEvent } from 'aws-lambda';
-import {
-  algoliaApiKey,
-  algoliaAppId,
-  algoliaIndex,
-  appName,
-  baseUrl,
-  contentfulAccessToken,
-  contentfulEnvId,
-  contentfulManagementAccessToken,
-  contentfulSpaceId,
-  isContentfulEnabledV2,
-} from '../../config';
+import { algoliaApiKey, algoliaAppId, algoliaIndex } from '../../config';
 import ExternalAuthors, {
   ExternalAuthorsController,
 } from '../../controllers/external-authors';
-import { ExternalAuthorContentfulDataProvider } from '../../data-providers/contentful/external-authors.data-provider';
-import {
-  ExternalAuthorDataProvider,
-  ExternalAuthorSquidexDataProvider,
-} from '../../data-providers/external-authors.data-provider';
-import { getAuthToken } from '../../utils/auth';
 import logger from '../../utils/logger';
 import { sentryWrapper } from '../../utils/sentry-wrapper';
 import { ExternalAuthorEvent, ExternalAuthorPayload } from '../event-bus';
+import { getExternalAuthorDataProvider } from '../../dependencies/external-authors.dependencies';
 
 export const indexExternalAuthorHandler =
   (
@@ -77,50 +52,10 @@ export const indexExternalAuthorHandler =
     }
   };
 
-let externalAuthorDataProvider: ExternalAuthorDataProvider;
-
-if (isContentfulEnabledV2) {
-  const contentfulGraphQLClient = getContentfulGraphQLClient({
-    space: contentfulSpaceId,
-    accessToken: contentfulAccessToken,
-    environment: contentfulEnvId,
-  });
-
-  const getContentfulRestClientFactory = () =>
-    getContentfulRestClient({
-      space: contentfulSpaceId,
-      accessToken: contentfulManagementAccessToken,
-      environment: contentfulEnvId,
-    });
-
-  externalAuthorDataProvider = new ExternalAuthorContentfulDataProvider(
-    contentfulGraphQLClient,
-    getContentfulRestClientFactory,
-  );
-} else {
-  const externalAuthorRestClient = new SquidexRest<RestExternalAuthor>(
-    getAuthToken,
-    'external-authors',
-    {
-      appName,
-      baseUrl,
-    },
-  );
-  const squidexGraphqlClient = new SquidexGraphql(getAuthToken, {
-    appName,
-    baseUrl,
-  });
-
-  externalAuthorDataProvider = new ExternalAuthorSquidexDataProvider(
-    externalAuthorRestClient,
-    squidexGraphqlClient,
-  );
-}
-
 /* istanbul ignore next */
 export const handler = sentryWrapper(
   indexExternalAuthorHandler(
-    new ExternalAuthors(externalAuthorDataProvider),
+    new ExternalAuthors(getExternalAuthorDataProvider()),
     algoliaSearchClientFactory({
       algoliaApiKey,
       algoliaAppId,
